refactor: 도메인명 수정 #203 - 도메인명 수정
