+++ be15_DevEagles_FE/src/features/items/components/BaseItemModal.vue
