+++ be15_DevEagles_FE/src/features/items/components/PrimaryRegistModal.vue
