+++ be15_DevEagles_FE/src/features/membership/components/MembershipEditModal.vue
@@ -96,7 +96,7 @@
 
 <script setup>
   import { computed, ref, onMounted } from 'vue';
-  import BaseItemModal from '@/features/item/components/BaseItemModal.vue';
+  import BaseItemModal from '@/features/items/components/BaseItemModal.vue';
   import BaseForm from '@/components/common/BaseForm.vue';
   import BaseButton from '@/components/common/BaseButton.vue';
   import MembershipDeleteModal from '@/features/membership/components/MembershipDeleteModal.vue';
