+++ be15_DevEagles_FE/src/routes/index.js
@@ -85,7 +85,7 @@ const routes = [
       {
         path: 'item/service',
         name: 'ItemService',
-        component: () => import('@/features/item/view/ItemView.vue'),
+        component: () => import('@/features/items/views/ItemView.vue'),
       },
       {
         path: 'item/membership',
